requirements.txt
aiohttp==3.9.3
python-dotenv==1.0.0
supabase==2.3.0
spacy
# For English model, user must run: python -m spacy download en_core_web_sm 